feat: package ready; add pyproject.toml, __init__.py, modify imports __init__.py: Expose bootstrapping, data_handling, and plotting as they are required for standard usage. Expose model_functions as well so the user can use them for custom plotting or analytical calculations.
pyproject.toml:
Build with hatchling.
Use dynamic versioning utilizing versioningit. The format for the versions
is obtained from the versioningit docs
https://versioningit.readthedocs.io/en/stable/ in revision 6c381341.
To build the package versioningit needs a tag which will be created for
the merge commit merging branch package_exoboots with main.

Python version:
Python in version 3.12 is required as exoboots uses the case
construction which was introduced in 3.10 and nested string literals in
format strings wich was introduced in 3.12.